2030 India Kidney Test Market Size, Growth Forecast, and Key Drivers.md
The India Kidney Function Test Market was estimated by USD 46.35 million in 2024. The market is predicted to grow to 70.94 millions by 2030 and growing at an average CAGR of 7.35 percent during the forecast time. The increasing incidence of chronic kidney diseases (CKD) as well as diabetes and hypertension are driving the the demand for regular kidney function tests in India. Additionally, the growing awareness of people about the early diagnosis and preventive health as well as improved health infrastructure in semi-urban and urban regions and a growing market. Initiatives by the government to promote the non-communicable diseases (NCD) examination and outreach to healthcare also aid in the expansion of this industry.