India Herbal Supplement Market Trends 2025 to 2030: Innovation Meets Tradition.md
India Herbal Supplement Market was estimated at USD 1.87 Billion in 2024. It is expected to grow steadily over the coming years with an estimated CAGR of 5.40 percent until 2030. The growing demand for natural remedies and products made from plants in healthcare, as well as growing awareness of prevention-based health practices is boosting need for supplements made from herbs. Consumers are increasingly inclined to herbs and Ayurvedic solutions to manage the effects of lifestyle on health, including obesity, diabetes, stress and arthritis. Additionally, the government's focus on encouraging Ayurveda through initiatives such as AYUSH and campaigns to support organic and local-sourced products is also promoting growth in the market.