2030 Forecast: Why India’s Vitamin D Supplements Market is Booming.md
The India Vitamin D Supplements Market was estimated at USD 98.02 Million in 2024. It is projected to grow to USD 157.81 Million by 2030 with an estimated CAGR of 8.22 percent during the forecast time. The market is experiencing steady growth, driven by the growing consciousness of vitamin D deficiencies and its health risks associated with it among the Indian population. With increasing urbanization and lifestyle changes, as well as the increasing amount of time spent indoors sun exposure has dramatically decreased, resulting in the widespread deficiency in vitamin D, which affects 70 to 90 percent of the population of India particularly those living in urban areas. The pandemic is also fueling interest in preventive health and immune-supportive supplements, resulting in a rising demands for vitamin D in capsules and soft gels, gummies and fortified foods.