2030 UAE Minimally Invasive Biopsy Techniques Market Forecast & Regional Analysis.md
The UAE Minimally Invasive Biopsy Techniques Market is experiencing steady growth due to the nation's advanced healthcare infrastructure and the increasing acceptance of non-invasive diagnostic methods. The market in 2024 was at USD 68.02 million and is expected to reach USD 82.65 million by 2030 increasing at a rate of 3.50 percent. The increasing incidence of cancer particularly prostate, breast and lung cancers is driving the demand for accurate and user-friendly diagnostic methods. The healthcare sector in the UAE is in the process of undergoing a massive digitization, which is accompanied by the use of minimally-invasive diagnostic techniques that provide less patient discomfort and quicker recovery times.